requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
ansicon==1.89.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
croniter==1.4.1
cycler==0.12.0
dateutils==0.6.12
deepdiff==6.4.1
docker-pycreds==0.4.0
easyocr==1.7.1
einops==0.2.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.9.0
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.0
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.5
imgaug==0.4.0
iniconfig==2.0.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.0.8
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
Pillow==9.3.0
pluggy==1.3.0
prefetch-generator==1.0.1
protobuf==4.24.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.8
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.13.3
tomli==2.0.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cu118
torch-tb-profiler==0.4.1
torchaudio==2.0.2+cu118
torchmetrics==1.1.1
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
vietocr==0.3.11
wandb==0.15.10
wcwidth==0.2.6
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
